Separate movie objects
Created a .py file where movie objects are stored for use on other packages
